[stdlib] Refactor: move static casing functions in native-wasm from generators to code
